+++ src/Film.ts
@@ -20,15 +20,6 @@ export interface Film {
   readonly id: string;
 
   /**
-   * Width and height within the source atlas image in integral pixels.
-   * Dimensions are identical for every cel.
-   */
-  readonly wh: Readonly<U16XY>;
-
-  /** Every film is expected to have at least one cel. */
-  readonly cels: readonly Cel[];
-
-  /**
    * Positive film length in milliseconds for a full cycle, possibly infinite.
    * For a ping-pong film, this is a full traversal forward plus the traversal
    * backward excluding the first and last frame. Eg, in a five cel animation,
@@ -38,6 +29,63 @@ export interface Film {
    * This is a U32, not a U16, since its an aggregation of U16s.
    */
   readonly duration: U32Millis | InfiniteDuration;
+
+  /**
+   * Width and height within the source atlas image in integral pixels.
+   * Dimensions are identical for every cel.
+   */
+  readonly wh: Readonly<U16XY>;
+
+  /** Every film is expected to have at least one cel. */
+  readonly cels: readonly Cel[];
+
+  /**
+   * Map each time division to a cel index. This is precomputed by the parser to
+   * save processing at the expense of memory.
+   *
+   * Eg, consider an animation with three cels:
+   *
+   *   Duration  Time Window   Cel Index
+   *   100 ms      0 -  99 ms  0
+   *   400 ms    100 - 499 ms  1
+   *   200 ms    500 - 699 ms  2
+   *
+   * The greatest common multiple duration is 100 ms so the frequency of
+   * divisions is 1 / 100 ms. The total duration is 700 ms.
+   *
+   *   number of divisions = frequency * total duration
+   *                       = (1 / 100 ms) * 700 ms
+   *                       = 7
+   *
+   * Each division maps to a cel index: [0, 1, 1, 1, 1, 2, 2]. A time is mapped
+   * to division and then to an index via: trunc(time * frequency).
+   *
+   * Infinite durations are ignored for multi-cel animation greatest common
+   * multiple computation. Single cel animations with infinite duration have a
+   * greatest common multiple of 1.
+   *
+   * The same example with an additional infinite duration cel:
+   *
+   *   Duration  Time Window   Cel Index
+   *   100 ms      0 -  99 ms  0
+   *   400 ms    100 - 499 ms  1
+   *   200 ms    500 - 699 ms  2
+   *     ∞ ms    700 -   ∞ ms  3
+   *
+   * Greatest common multiple: 100 ms
+   * Divisions: [0, 1, 1, 1, 1, 2, 2, 3]
+   *
+   * The cels property could be duplicated by reference and indexed directly but
+   * this format is intended to be JSON friendly.
+   */
+  readonly celIndexByDivision: readonly U16[];
+
+  /**
+   * The width of each division, possibly infinite. This is the greatest common
+   * multiple of cel durations, excluding infinite durations. See
+   * celIndexByDivision.
+   */
+  readonly timeDivision: U32Millis | InfiniteDuration;
 
   readonly direction: Playback;
 }